requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
openai>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
moviepy>=1.0.3
openai-whisper>=20231117
segment-anything-py>=0.1.0
pillow>=10.0.0
tqdm>=4.65.0
python-ffmpeg>=2.0.4 